chore: new hooks
