--- packages/runtime/hooks/.babelrc
@@ -1,9 +1,0 @@
-{
-	"presets": [
-		["@babel/preset-env", {
-			"targets": {
-				"node": "current"
-			}
-		}]
-	]
-}
