--- packages/runtime/hooks/.editorconfig
@@ -1,16 +1,0 @@
-root = true
-
-[*]
-indent_style = tab
-indent_size = 2
-charset = utf-8
-trim_trailing_whitespace = true
-insert_final_newline = true
-max_line_length = 80
-
-[*.{yml,yaml,json}]
-indent_style = space
-indent_size = 2
-
-[*.md]
-trim_trailing_whitespace = false
