--- packages/runtime/hooks/.gitattributes
@@ -1,1 +1,0 @@
-* text=auto
