--- packages/runtime/hooks/.gitignore
@@ -1,12 +1,0 @@
-/node_modules
-/coverage
-
-############
-## Windows
-############
-
-# Windows image file caches
-Thumbs.db
-
-# Folder config file
-Desktop.ini
