--- packages/runtime/hooks/.prettierrc.js
@@ -1,13 +1,0 @@
-module.exports = {
-	printWidth: 80,
-	useTabs: true,
-	tabWidth: 2,
-	overrides: [
-		{
-			files: "*.json",
-			options: {
-				useTabs: false,
-			},
-		},
-	],
-};
