--- packages/runtime/hooks/.travis.yml
@@ -1,9 +1,0 @@
-sudo: false
-language: node_js
-node_js:
-  - "12"
-  - "10"
-  - "8"
-  - "6"
-
-script: yarn run travis
