+++ packages/runtime/hooks/src/source/SyncLoopHook.js
@@ -1,0 +1,46 @@
+/*
+	MIT License http://www.opensource.org/licenses/mit-license.php
+	Author Tobias Koppers @sokra
+*/
+'use strict';
+
+const Hook = require('./Hook');
+const HookCodeFactory = require('./HookCodeFactory');
+
+class SyncLoopHookCodeFactory extends HookCodeFactory {
+  content({ onError, onDone, rethrowIfPossible }) {
+    return this.callTapsLooping({
+      onError: (i, err) => onError(err),
+      onDone,
+      rethrowIfPossible,
+    });
+  }
+}
+
+const factory = new SyncLoopHookCodeFactory();
+
+const TAP_ASYNC = () => {
+  throw new Error('tapAsync is not supported on a SyncLoopHook');
+};
+
+const TAP_PROMISE = () => {
+  throw new Error('tapPromise is not supported on a SyncLoopHook');
+};
+
+const COMPILE = function (options) {
+  factory.setup(this, options);
+  return factory.create(options);
+};
+
+function SyncLoopHook(args = [], name = undefined) {
+  const hook = new Hook(args, name);
+  hook.constructor = SyncLoopHook;
+  hook.tapAsync = TAP_ASYNC;
+  hook.tapPromise = TAP_PROMISE;
+  hook.compile = COMPILE;
+  return hook;
+}
+
+SyncLoopHook.prototype = null;
+
+module.exports = SyncLoopHook;
