+++ packages/runtime/hooks/src/source/index.js
@@ -1,11 +1,11 @@
 // fork from https://github.com/webpack/tapable
 
-export { SyncHook } from "./SyncHook";
-export { SyncBailHook } from "./SyncHook";
-export { AsyncParallelHook } from "./AsyncParallelHook";
-export { AsyncParallelBailHook } from "./AsyncParallelBailHook";
-export { AsyncSeriesHook } from "./AsyncSeriesHook";
-export { AsyncSeriesBailHook } from "./AsyncSeriesBailHook";
+export { SyncHook } from './SyncHook';
+export { SyncBailHook } from './SyncHook';
+export { AsyncParallelHook } from './AsyncParallelHook';
+export { AsyncParallelBailHook } from './AsyncParallelBailHook';
+export { AsyncSeriesHook } from './AsyncSeriesHook';
+export { AsyncSeriesBailHook } from './AsyncSeriesBailHook';
 
 // exports.SyncWaterfallHook = require('./SyncWaterfallHook');
 // exports.SyncLoopHook = require('./SyncLoopHook');
